requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.2
djangorestframework==3.14.0
imutils==0.5.4
numpy==1.26.4
opencv-python==4.9.0.80
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
